bpf: Make the XDP mb and mb_frag accessible trough the ctx.
This will avoid calling a helper just to see if a mb is
present, so no registers need to be saved on the stack, etc.

Not sure if mb_frag is needed, but it might be useful so the
parameter does not have to be passed down trough functions.

Also though of making mb available as the number of
fragments available, i.e. mb_nr_frags. However, getting a
pointer to skb_shared_info from eBPF code without two scratch
register requires some more work, but is possible.

Signed-off-by: Eelco Chaudron <echaudro@redhat.com>